WHO Vaccine Visualization in Power BI Service
Portfolio project to load, wrangle, and visualize real COVID19 vaccination data from the WHO: Visualization component. Screen recording available.

About this visualization

This is is basic, functional visualization created by Vickie Gray for a personal portfolio. The data is real global COVID 19 Vaccination data downloaded and wrangled from the World Health Organization (WHO).

The visualization takes into account visual accessibility and efficiency, and is clean, simple and adaptable. It includes a drill-down matrix of WHO Region and country vaccination data about vaccinations per 100 people, a WHO Region slicer, the stacked bar graph of total vaccinations includes a zoom slider to visualize country statistics at smaller scales, and a tooltip to explain the WHO Region codes. This is an example of a basic visualization I would bring to a client for further customization according to their requirements.

This visualization is built in a trial of the Power BI Service, from a .csv file downloaded from the World Health Organization and loaded into Excel 365 for wrangling. It is a static data file, from the second week of February, 2023. I chose to use a static file, as I only have a 45-day trial of the Power BI Service. Normally, I would use the desktop client and evergreen data.

A Jupyter Notebook is available in this folder that analyzes and discusses this same data in Python, Pandas, Matplotlib, and Seaborn.